requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
idna==3.3
importlib-metadata==4.11.3
jeepney==0.7.1
jira==3.1.1
keyring==23.5.0
oauthlib==3.2.0
pycparser==2.21
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
urllib3==1.26.9
zipp==3.7.0
